+++ NEWS.md
@@ -4,6 +4,9 @@ SHELF v1.5.0.9000 (2019-04-02)
 * elicitMultiple() app: can now enter judgements with the roulette method, and save/load judgements 
 as .csv files
 
+* column names changed in output of feedback(), fitdist() and sampleFit() to be consistent:
+"normal", "t", "gamma", "lognormal", "logt", "beta", "hist"
+
 
 SHELF v1.5.0 (2019-03-26) 
 ==============================